+++ src/main.rs
@@ -2,11 +2,14 @@
 #![no_main]
 #![allow(clippy::unusual_byte_groupings, clippy::deref_addrof)]
 #![forbid(static_mut_refs)]
-#![feature(abi_x86_interrupt)]
+#![feature(abi_x86_interrupt, sync_unsafe_cell)]
 
 /// Allows writing to the VGA text buffer
 #[macro_use]
 mod vga;
+
+/// Handles the InitLater and UnsafeFlag wrappers.
+mod wrappers;
 
 /// Handles various interrupts
 mod interrupts;
@@ -29,8 +32,9 @@ mod time;
 // Warn anyone just running `cargo build` to use the bootimage tool
 #[cfg(any(debug_assertions, not(feature = "bootimage")))]
 compile_error!(
-    "Please build sunflower using `cargo b` or `cargo bootimage` and 
-run it using either `run-nosound` `run-pipewire` or `run-pulseaudio`"
+    "Please build sunflower using `cargo b` or `cargo bootimage` 
+run it using `cargo run-nosound` `cargo run-pipewire` or `cargo run-pulseaudio` 
+and use clippy via `cargo paperclip`"
 );
 
 /// The kernel entry point.
@@ -41,12 +45,12 @@ pub unsafe extern "C" fn kmain() -> ! {
     startup::run("Connected VGA", vga::init);
     startup::run("Loaded IDT", interrupts::load_idt);
     startup::run("Initialised PIC", interrupts::init_pic);
+    startup::run("Prepared RTC sync", time::setup_rtc_int);
     startup::run("Set PIT frequency", time::set_timer_interval);
     startup::run("Initialised keyboard", interrupts::init_kbd);
-    startup::run("Prepared for RTC sync", time::setup_rtc_int);
     startup::run("Checked CPUID", sysinfo::check_cpuid);
+    startup::run("Finished RTC sync", time::wait_for_rtc_sync);
 
-    startup::SYS_INIT.store(true, core::sync::atomic::Ordering::Relaxed);
     vga::print_color("All startup tasks completed! \u{1}\n\n", vga::Color::Green);
     vga::update_vga_cursor();
     speaker::play_chime();